+++ MVP.xcodeproj/xcuserdata/whanna.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,6 +4,16 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
+		<key>MVP (Development).xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>6</integer>
+		</dict>
+		<key>MVP (Production).xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>5</integer>
+		</dict>
 		<key>MVP.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>isShown</key>
