+++ MVP/Configs/Development.xcconfig
@@ -1,0 +1,17 @@
+//
+//  Development.xcconfig
+//  MVP
+//
+//  Created by Wasim Hanna on 10/16/19.
+//  Copyright © 2019 Insight. All rights reserved.
+//
+
+// Configuration settings file format documentation can be found at:
+// https://help.apple.com/xcode/#/dev745c5c974
+
+#include "Pods/Target Support Files/Pods-MVP/Pods-MVP.debug (development.xcconfig"
+
+HTTP_SCHEME = https
+BASE_URL = insightmvp-dev.azurewebsites.net
+SCORE_URL = /api/Matches/recent
+LEADERBOARD = /api/Ratings/topPlayers
