+++ MVP/Configs/Production.xcconfig
@@ -1,0 +1,16 @@
+//
+//  Production.xcconfig
+//  MVP
+//
+//  Created by Wasim Hanna on 10/16/19.
+//  Copyright © 2019 Insight. All rights reserved.
+//
+
+// Configuration settings file format documentation can be found at:
+// https://help.apple.com/xcode/#/dev745c5c974
+#include "Pods/Target Support Files/Pods-MVP/Pods-MVP.release (production).xcconfig"
+
+HTTP_SCHEME = https
+BASE_URL = insightmvp.azurewebsites.net
+SCORE_URL = /api/Matches/recent
+LEADERBOARD = /api/Ratings/topPlayers
